first-fastapi/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.101.0
h11==0.14.0
idna==3.4
pydantic~=1.10.11
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
tortoise-orm==0.19.3
asyncpg==0.28.0